Remove unnecessary inclusion of agent class
Prior to this comment, the puppet::agent class was automatically included by the
config class. As a result, any resource-style declaration of puppet::agent would
cause a duplicate resource failure. Since the puppet::agent class does not
contain variables used by puppet::config, the inclusion is removed in this
commit.